pyproject.toml
[tool.poetry]
name = "fastapi-oauth-demo"
version = "0.1.0"
description = ""
authors = ["t-okuji <t.okuji34@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic-settings = "^2.4.0"
requests = "^2.32.3"
pyjwt = "^2.9.0"
cryptography = "^43.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pre-commit = "^3.7.1"
debugpy = "^1.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ruff configure
[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
]

ignore = [
    "D105", # undocumented-magic-method
    "D107", # undocumented-pubic-init
    "D405", # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
known-third-party = ["fastapi"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
docstring-code-format = false
docstring-code-line-length = "dynamic"
